+++ tvshows/src/components/features/review/ReviewItem/ReviewItem.tsx
@@ -1,6 +1,6 @@
 import { IReview } from '@/typings/review.types';
 import { Box, Button, Flex, Image, Text } from '@chakra-ui/react';
-import StarRating from './StarRating';
+import StarRating from '../StarRating/StarRating';
 
 const maxRating = '5';
 
@@ -48,7 +48,7 @@ export const ReviewItem = ({ reviewItem, onDelete }: IReviewItemProps) => {
           : 'No rating'}
       </Box> */}
       <Flex align='center' flexDirection='row' mt={5} ml={5}>
-        <StarRating defaultValue={reviewItem.rating} onChange={() => {}} mode={1} />
+        <StarRating defaultValue={reviewItem.rating} onChange={() => {}} mode={'static'} />
         <Text ml={5} mt={1} as='b'>
           {reviewItem.rating
             ? `${reviewItem.rating} / ${maxRating}`
